Starting information and folder
Created a couple informational documents to help the project progress smoothly. Also created a few batch scripts to help with compiling and running the software after we eventually create it. These batch scripts still need to be thoroughly tested and debugged when we have .java files to compile and run